Fixed multiple notes with same name clashing Cleanup of notes.qml code Added todo move unchecked items up Added todo move checked items down Added visual effects for checked and unchecked notes

+++ qml/config.js
@@ -5,6 +5,19 @@ function getDatabase() {
     return LS.LocalStorage.openDatabaseSync("Noto", "1.0", "StorageDatabase", 100000);
 }
 
+// We want a unique id for notes
+function getUniqueId()
+{
+     var dateObject = new Date();
+     var uniqueId =
+          dateObject.getFullYear() + '' +
+          dateObject.getMonth() + '' +
+          dateObject.getDate() + '' +
+          dateObject.getTime();
+
+     return uniqueId;
+};
+
 // At the start of the application, we can initialize the tables we need if they haven't been created yet
 function initialize() {
     var db = getDatabase();
@@ -12,21 +25,21 @@ function initialize() {
                 function(tx) {
                     // Create the settings table if it doesn't already exist
                     // If the table exists, this is skipped
-                    tx.executeSql('CREATE TABLE IF NOT EXISTS notes(title TEXT UNIQUE,txt TEXT)');
+                    tx.executeSql('CREATE TABLE IF NOT EXISTS notes(uid INTEGER UNIQUE, title TEXT,txt TEXT)');
+                    // TODO: Add uid for title, todo
                     tx.executeSql('CREATE TABLE IF NOT EXISTS todos(title TEXT,todo TEXT,status INTEGER)');
                     tx.executeSql('CREATE UNIQUE INDEX IF NOT EXISTS idx ON todos(title,todo);');
                 });
 }
 
 // This function is used to write notes into the database
-function setNote(title,txt) {
+function setNote(uid,title,txt) {
     // title: name representing the title of the note
     // txt: text of the note
     var db = getDatabase();
     var res = "";
     db.transaction(function(tx) {
-        var rs = tx.executeSql('INSERT OR REPLACE INTO notes VALUES (?,?);', [title,txt]);
-        //console.log(rs.rowsAffected)
+        var rs = tx.executeSql('INSERT OR REPLACE INTO notes VALUES (?,?,?);', [uid,title,txt]);
         if (rs.rowsAffected > 0) {
             res = "OK";
             console.log ("Saved to database");
@@ -69,9 +82,10 @@ function getNotes() {
     var db = getDatabase();
     var respath="";
     db.transaction(function(tx) {
-        var rs = tx.executeSql('SELECT DISTINCT title FROM notes;');
+        var rs = tx.executeSql('SELECT DISTINCT title,uid FROM notes;');
         for (var i = 0; i < rs.rows.length; i++) {
-            root.addNote(rs.rows.item(i).title)
+            root.addNote(rs.rows.item(i).title,rs.rows.item(i).uid)
+            console.debug("get Notes:" + rs.rows.item(i).title + " with id:" + rs.rows.item(i).uid)
         }
     })
 }
@@ -89,11 +103,12 @@ function getTodos() {
 }
 
 // This function is used to retrieve a text from a note in the database
-function getText(title) {
+function getText(title,uid) {
     var db = getDatabase();
     var notesText="";
     db.transaction(function(tx) {
-        var rs = tx.executeSql('SELECT txt FROM notes WHERE title=?;', [title]);
+        var rs = tx.executeSql('SELECT txt FROM notes WHERE title=? AND uid=?;', [title,uid]);
+        console.debug(title + " " + uid)
         if (rs.rows.length > 0) {
             notesText = rs.rows.item(0).txt
         } else {
@@ -115,8 +130,8 @@ function getTodo(title) {
         for (var i = 0; i < rs.rows.length; i++) {
             if (rs.rows.length > 0) {
                 var row = rs.rows.item(i)
-                console.debug(row['todo'])
-                console.debug(row['status'])
+                //console.debug(row['todo'])
+                //console.debug(row['status'])
                 todoPage.addTodo(row['todo'],row['status'])
             } else {
                 todoText = "Unknown"
@@ -130,12 +145,13 @@ function getTodo(title) {
 }
 
 // This function is used to remove a note or todo from the database
-function remove(title,type) {
+function remove(title,type,uid) {
     var db = getDatabase();
     var respath="";
     if (type === "note") {
+        console.debug("Removing Note: " + uid)
         db.transaction(function(tx) {
-            var rs = tx.executeSql('DELETE FROM notes WHERE title=?;' , [title]);
+            var rs = tx.executeSql('DELETE FROM notes WHERE title=? AND uid=?;' , [title,uid]);
         })
     }
     else if  (type === "todo") {
